Add workers seed
